refactor: move the navbar to its own component
